Moved the default generator to a new one: 'enumerate_it:enum'. Also created an install task: 'enumerate_it:install' that copies the default initializer file to the config/initializers folder of a Rails project.
