+++ lib/generators/enumerate_it/enum/templates/enumerate_it.rb
