+++ lib/generators/enumerate_it/enum/templates/locale.yml
