+++ lib/generators/enumerate_it/install/USAGE
@@ -1,0 +1,8 @@
+Description:
+    Creates an initializer file for EnumerateIt
+
+Example:
+    rails generate enumerate_it:install
+
+    This will create:
+        config/initializers/enumerate_it.rb
