+++ lib/generators/enumerate_it/install/install_generator.rb
@@ -1,0 +1,11 @@
+module EnumerateIt
+  module Generators
+    class InstallGenerator < Rails::Generators::Base
+      source_root File.expand_path('../templates', __FILE__)
+
+      def copy_initializer_file
+        template "enumerate_it_initializer.rb", File.join('config/initializers/', "enumerate_it.rb")
+      end
+    end
+  end
+end
