+++ lib/generators/enumerate_it/install/templates/enumerate_it_initializer.rb
@@ -1,0 +1,1 @@
+ActiveRecord::Base.extend EnumerateIt
